PI: Overall improvements

+++ klibc/stdio/feof.c
@@ -4,7 +4,7 @@
 /*@
 	requires valid_FILE(__f);
 	assigns \nothing;
-	ensures \result == __f->_IO_eof;
+	ensures \result == (int)(__f->_IO_eof);
 @*/
 int feof(FILE *__f)
 {
